Moves manual price logic to the new component
