+++ frontend/app/src/components/accounts/manual-balances/ManualBalancesForm.vue
@@ -10,9 +10,9 @@
       class="manual-balances-form__label"
       outlined
       :label="tc('manual_balances_form.fields.label')"
-      :error-messages="v$.label.$errors.map(e => e.$message)"
+      :error-messages="toMessages(v$.label.$errors)"
       :disabled="pending"
-      @focus="delete errors['label']"
+      @blur="v$.label.$touch()"
     />
 
     <balance-type-input
@@ -39,9 +39,9 @@
       :label="tc('common.asset')"
       class="manual-balances-form__asset"
       outlined
-      :error-messages="v$.asset.$errors.map(e => get(e.$message))"
+      :error-messages="toMessages(v$.asset.$errors)"
       :disabled="pending"
-      @focus="delete errors['asset']"
+      @blur="v$.asset.$touch()"
     />
 
     <v-row v-else class="mb-n9">
@@ -52,7 +52,7 @@
           persistent-hint
           clearable
           :disabled="pending"
-          :error-messages="v$.customAssetName.$errors.map(e => e.$message)"
+          :error-messages="toMessages(v$.customAssetName.$errors)"
           :label="t('common.name')"
         />
       </v-col>
@@ -66,35 +66,29 @@
           clearable
           :disabled="pending"
           :label="t('common.type')"
-          :error-messages="v$.customAssetType.$errors.map(e => e.$message)"
+          :error-messages="toMessages(v$.customAssetType.$errors)"
           :search-input.sync="search"
         />
       </v-col>
     </v-row>
 
     <manual-balances-price-form
-      :price="price"
-      :price-asset="priceAsset"
-      :fetched-price="fetchedPrice"
-      :fetching-price="fetchingPrice"
-      :is-custom-price="isCustomPrice"
+      ref="priceForm"
       :pending="pending"
       :asset-method="assetMethod"
-      @update:price="price = $event"
-      @update:price-asset="priceAsset = $event"
-      @update:custom-price="isCustomPrice = $event"
     />
 
     <amount-input
       v-model="amount"
       :label="tc('common.amount')"
-      :error-messages="errors['amount']"
+      :error-messages="toMessages(v$.amount.$errors)"
       class="manual-balances-form__amount"
       outlined
       autocomplete="off"
       :disabled="pending"
-      @focus="delete errors['amount']"
+      @blur="v$.amount.$touch()"
     />
+
     <tag-input
       v-model="tags"
       :label="tc('manual_balances_form.fields.tags')"
@@ -102,14 +96,15 @@
       outlined
       class="manual-balances-form__tags"
     />
+
     <location-selector
       v-model="location"
       class="manual-balances-form__location"
       outlined
-      :error-messages="errors['location']"
+      :error-messages="toMessages(v$.location.$errors)"
       :disabled="pending"
       :label="tc('common.location')"
-      @focus="delete errors['location']"
+      @blur="v$.location.$touch()"
     />
   </v-form>
 </template>
@@ -117,7 +112,6 @@
 <script setup lang="ts">
 import useVuelidate from '@vuelidate/core';
 import { helpers, required, requiredIf } from '@vuelidate/validators';
-import { get, set } from '@vueuse/core';
 import { PropType, Ref } from 'vue';
 import ManualBalancesPriceForm from '@/components/accounts/manual-balances/ManualBalancesPriceForm.vue';
 import LocationSelector from '@/components/helper/LocationSelector.vue';
@@ -130,14 +124,12 @@ import { deserializeApiErrorMessage } from '@/services/converters';
 import { api } from '@/services/rotkehlchen-api';
 import { useBalancesStore } from '@/store/balances';
 import { useManualBalancesStore } from '@/store/balances/manual';
-import { useBalancePricesStore } from '@/store/balances/prices';
 import { useMessageStore } from '@/store/message';
-import { useGeneralSettingsStore } from '@/store/settings/general';
-import { CURRENCY_USD } from '@/types/currencies';
 import { TradeLocation } from '@/types/history/trade-location';
 import { ManualBalance } from '@/types/manual-balances';
 import { startPromise } from '@/utils';
 import { bigNumberify } from '@/utils/bignumbers';
+import { toMessages } from '@/utils/validation-errors';
 
 const props = defineProps({
   edit: {
@@ -173,11 +165,8 @@ const balanceType: Ref<BalanceType> = ref(BalanceType.ASSET);
 const form = ref<any>(null);
 const customAssetTypes = ref<string[]>([]);
 const search = ref<string | null>('');
-const price = ref<string>('');
-const priceAsset = ref<string>('');
-const fetchingPrice = ref<boolean>(false);
-const fetchedPrice = ref<string>('');
-const isCustomPrice = ref<boolean>(false);
+const priceForm: Ref<InstanceType<typeof ManualBalancesPriceForm> | null> =
+  ref(null);
 
 watch(search, search => {
   if (search === null) search = '';
@@ -224,8 +213,6 @@ watch(
 const { editManualBalance, addManualBalance, manualLabels } =
   useManualBalancesStore();
 const { refreshPrices } = useBalancesStore();
-const { fetchPrices, toSelectedCurrency, getAssetPrice } =
-  useBalancePricesStore();
 const { setMessage } = useMessageStore();
 
 const saveCustomAsset = async (): Promise<string | undefined> => {
@@ -276,13 +263,8 @@ const save = async () => {
     : addManualBalance({ ...balance, asset: usedAsset }));
 
   if (status.success) {
-    if (get(isCustomPrice) && get(price) && get(priceAsset)) {
-      await api.assets.addLatestPrice({
-        fromAsset: usedAsset,
-        toAsset: get(priceAsset),
-        price: get(price)
-      });
-    }
+    const form = get(priceForm);
+    await form?.savePrice(usedAsset);
   }
 
   set(pending, false);
@@ -367,6 +349,9 @@ const rules = {
       tc('asset_form.type_non_empty'),
       requiredIf(() => get(assetMethod) === 1)
     )
+  },
+  location: {
+    required
   }
 };
 
@@ -376,6 +361,7 @@ const v$ = useVuelidate(
     amount,
     asset,
     label,
+    location,
     customAssetName,
     customAssetType
   },
@@ -388,74 +374,23 @@ watch(v$, ({ $invalid }) => {
 
 watch(valid, value => input(value));
 
-defineExpose({
-  save
+watch(asset, async asset => {
+  const form = get(priceForm);
+  await form?.searchAssetPrice(asset);
 });
-
-const { currencySymbol } = storeToRefs(useGeneralSettingsStore());
-
-const searchAssetPrice = async (asset: string) => {
-  if (!asset) {
-    set(price, '');
-    set(priceAsset, '');
-    set(fetchedPrice, '');
-    set(isCustomPrice, true);
-    return;
-  }
-
-  set(fetchingPrice, true);
-  await fetchPrices({
-    ignoreCache: false,
-    selectedAssets: [asset]
-  });
-  set(fetchingPrice, false);
-
-  const priceInUsd = getAssetPrice(asset);
-  if (priceInUsd && !priceInUsd.eq(0)) {
-    const priceInCurrentRate = get(toSelectedCurrency(priceInUsd)).toFixed();
-    set(price, priceInCurrentRate);
-    set(priceAsset, get(currencySymbol));
-    set(fetchedPrice, priceInCurrentRate);
-    set(isCustomPrice, false);
-  } else {
-    set(isCustomPrice, true);
-  }
-};
 
 onMounted(async () => {
   set(customAssetTypes, await api.assets.getCustomAssetTypes());
 
   const editPayload = get(edit);
   if (editPayload) {
-    const asset = editPayload.asset;
-    await searchAssetPrice(asset);
+    const form = get(priceForm);
+    await form?.searchAssetPrice(editPayload.asset);
   }
 });
 
-watch(isCustomPrice, isCustomPrice => {
-  if (isCustomPrice) {
-    set(price, '');
-    set(priceAsset, '');
-  } else {
-    set(price, get(fetchedPrice));
-    set(priceAsset, get(currencySymbol));
-  }
-});
-
-watch(assetMethod, assetMethod => {
-  if (assetMethod === 1) {
-    set(price, '');
-    set(priceAsset, '');
-    set(isCustomPrice, true);
-  } else if (get(fetchedPrice)) {
-    set(price, get(fetchedPrice));
-    set(priceAsset, CURRENCY_USD);
-    set(isCustomPrice, false);
-  }
-});
-
-watch(asset, async asset => {
-  await searchAssetPrice(asset);
+defineExpose({
+  save
 });
 </script>
